init commit

+++ server/src/models.rs
@@ -1,16 +1,15 @@
+use crate::schema::posts;
 use diesel::prelude::*;
 
 #[derive(Queryable, Selectable)]
-#[diesel(table_name = crate::schema::posts)]
+#[diesel(table_name = posts)]
 #[diesel(check_for_backend(diesel::mysql::Mysql))]
 pub struct Post {
-    pub id: i64,
+    pub id: i32,
     pub title: String,
     pub body: String,
     pub published: bool,
 }
-
-use crate::schema::posts;
 
 #[derive(Insertable)]
 #[diesel(table_name = posts)]
@@ -18,3 +17,5 @@ pub struct NewPost<'a> {
     pub title: &'a str,
     pub body: &'a str,
 }
+
+
